perf: torch.compile graph break 대폭 제거 (68→~22)
- _Int8LinearCudaFn에 @torch.compiler.allow_in_graph 추가
- Int8LinearCuda.forward에서 @torch.compiler.disable 제거
- replace_int8linear_with_cuda() 함수 추가 (--int8_qat 시 자동 교체)
- Mamba2Block._forward_cuda를 forward + _fused_scan으로 분할
  (in_proj는 compiled graph에 남기고, fused kernel만 @disable)
- RMSNorm에 use_triton 플래그 추가 (--compile 시 자동 비활성화)
- DDP: --compile 시 find_unused_parameters=False로 전환
- tag_head 입력에서 불필요한 .float() 캐스팅 제거

+++ model/dense_editor.py
@@ -126,7 +126,8 @@ class DenseEditor(nn.Module):
                 x = layer(x, pad_mask=pad_mask, reset_mask=reset_mask)
 
         x = self.final_norm(x)
-        tag_logits = self.tag_head(x.float())
+        # tag_head 내부에서 LayerNorm + quantization 수행 → float 캐스팅은 CE loss 직전으로 이동
+        tag_logits = self.tag_head(x)
 
         return tag_logits
